requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
google-cloud-vision>=3.10,<4.0
python-multipart>=0.0.9,<1.0
pytesseract>=0.3.10,<1.0
pillow>=10.0,<12.0
matplotlib>=3.8,<4.0
